Provide ability to specify crypto mode
Bracket supports GCM and XTS encryption mode for the disks. By default
the guest root volume always was encrypted using GCM. This change
adds a new (hidden) --boot-disk-crypto-policy argument which allows
the user to override the default encryption policy for the (guest)
root volume. If 'default' is specified, then the default encryption
policy is determined from the encryptor image type (NetBSD vs FreeBSD).
Note that XTS is not supported with the NetBSD encryptor.