Use correct attribute to identify encryptor type for GCE
For GCE, it is the `image_name` which determines the encryptor
type (NetBSD vs FreeBSD). However the `image_name` would vary
depending on how it was determined:
 - User provided image file: whatever the user provided
 - Published images in public buckets: the corresponding image name
 - Default latest image: always returned as `latest.tar.gz`

The last case above is the most common use case where users will
simply used the `released` latest image to encrypt their GCE image.
For this case, the boot volume crypto policy would be set to
`default` allowing the encryptor to determine the policy itself.
In all other case, the crypto policy will be set based on the image
name.