Refactor make-user-data
Update the make-user-data command, so that it's a bit easier to test.
The work is now done by make_user_data.make() instead of the run()
method of the subcommand.

This change will be needed by upcoming change to all commands that print
formatted output to stdout.  make-user-data, make-key, make-token, and
get-public-key will all need a new option to print content to a file, so
that it doesn't get mangled by file redirection on Windows.