Add option to write to file instead of stdout
Add an --out option to the make-key, get-public-key, make-token, and
make-user-data commands, which causes output to be written to a file
instead of stdout.

This option is necessary on Windows, because PowerShell writes
redirected output as UTF-16.  Python is unable to make sense of UTF-16.
   For example, if you redirect the output of make-key to a file,
make-token will not recognize the file format.

Add a new argutil module, which will contain common command line
options.  Add a new function to util for writing output to either a file
or stdout.