Update front page and AWS docs
Update the description to include VMware, keys, and tokens.

Update links, so that they point to our provider-specific docs.  Chances
are that the reader already knows what AWS, GCE, and VMware are.

Move networking requirements to the AWS page, since they differ by CSP.
We will add networking requirements for GCE and VMware in a subsequent
commit.

Rearrange the AWS page so that the usage sections are at the bottom.
They're just a reference.  The shorter, descriptive sections should be
higher up.