Add a "/" to the metavisor directory in the ovf bucket
Since we're naming our artifacts <version> and <version>-debug we
were getting conflict when just calling bucket.list(<artifact name>).
In the case of debug it would work fine, but for release both debug
wand nondebug artifacts were downloaded.

This would have worked fine if it wasn't for the fact that one of the
files in both buckets is called disk-0.vmdk.