Add auth command
Add a top-level auth command that authenticates with the Bracket service
and prints an API auth token.  This command can be used by developers
and scripts that need to make calls to the Bracket REST API.  Update the
existing token validation code to use the same yeti module.

The next step will be to add login/logout commands (probably under
config), which use the yeti module to authenticate and stores the API
token in the user's config file.

---------------

$ ./brkt auth -h
usage: brkt auth [-h] [--email ADDRESS] [--password PASSWORD]
[--root-url URL]
                 [--out PATH]

Authenticate with the Bracket service. On success, print the JSON Web
Token
that can be used to make calls to Bracket REST API endpoints.

optional arguments:
--email ADDRESS      If not specified, show a prompt.
--out PATH           Write to a file instead of stdout. This can be
used to
                       avoid character encoding issues when redirecting
output
                       on Windows.
--password PASSWORD  If not specified, show a prompt.
--root-url URL       Bracket service root URL (default:
                       https://api.mgmt.brkt.com)
-h, --help           show this help message and exit