Clean up endpoint parsing
Clean up and unit test the code that parses an endpoint string.  Update
the function to return a tuple of (host, port) instead of a dictionary.
This is more intuitive, since the function will always return exactly
two values.  Update the parsing logic to be more rigorous about catching
edge cases.  Raise ValidationError in all cases where the string is
invalid, so that behavior is consistent.  Add unit tests.