Download artifacts from S3 only if required
Previously the brkt-cli would always try to download the Metavisor
artifacts from S3 everytime an encrypt/update command is run, even
if the command fails subsequently do to issues with the vCenter/ESX
server and/or the specified parameters. With this change, we:
 - Check if artifacts were already downloaded
 - Check if the VMDK hash matches the manifest file
Only when one or more of the files are not found or the manifest hash
does not match the VMDK, do we try to download the artifacts from S3.

Additionally a hidden --no-cleanup option has been provided which can
be used in case the user does not want the downloaded artifacts to be
deleted after the encryption/update completes successfully. This is
especially useful in the case where the same artifact is required to
encrypt/update multiple images and eliminates the need to download
the artifacts from S3 for each of the encrypt/update operations.