Check supported versions only once per day
Check whether the version of brkt-cli is supported once a day, instead
of checking every time the command runs.  This avoids an unnecessary
network round trip.  It also makes it less likely that users will see
warnings in the command output when the PyPI REST API is not responding.

Add support for internal options to CLIConfig.  These options are used
by brkt-cli application code and not exposed to end users.  Store the
last version check timestamp as an internal option.

Move version-related code to a new module, to make the main brkt_cli
module more concise.

Add the ability to read CLIConfig from any file, instead of hardcoding
CONFIG_PATH.  This allows us to write unit tests that use StringIO.