Update make-token to get token from the Bracket service
When the user doesn't specify a signing key, get a launch token from the
Bracket service.  Add a new get_token() method to YetiService.

Add a new --brkt-tag option for specifying JWT claims.  This is
effectively the same thing as --claim, but with more validation of the
key and value.  We will deprecate --claim in a future commit.