Get API token from an environment variable
Update config.get_yeti_service() to get the API token from the
BRKT_API_TOKEN environment variable instead of the user's config.
Turned out that storing the API token in the config file isn't
convenient for scripting.  With an environment variable, the workflow is
the same for end users and scripts.

Update brkt auth to read the default API URL from config instead of
hardcoding to the production service.