Get launch token from Bracket service during encryption
Update the encrypt and update commands to get the Metavisor launch token
from the Bracket service if it's not specified on the command line.

With this change, the CLI will always send a launch token to Metavisor
when encrypting or updating.  If the launch token is not specified with
the --token option, we will now require that the $BRKT_API_TOKEN
environment variable be set.  This allows the CLI to authenticate and
get a launch token from the Bracket service.