Create guest image snapshot when lacking permission
The wrap-image command (for AWS) made an assumption that the user
will have permission to launch the guest image snapshot. This is
true if the image is owned by the account where command is being
run. However in some cases, especially with the marketplace images,
the user might not have "create volume" permissions on the guest
root image snapshot. In such cases, the wrap command should:
	* Launch the guest image
	* Stop the guest instace
	* Snapshot the guest root volume
	* Terminate guest instance
	* Launch the created snapshot wrapped with the Metavisor
          image