Back up the old template till the new template is created
During update, the old template gets replaced with the new one.
During the transition, the old one gets deleted and and the new
one is added. If something goes wrong during the addition of the
new template, the old template is lost while the new one is not
created either. So, backup the old template, then create a new one,
and finally delete the old template.