Share-logs AWS Rewrite
This program creates a snapshot of the metavisors root disk,
attaches it to a volume, and launches a new instance
with the volume attached. The logs are then compressed
into a tar file and uploaded to a s3 bucket. This replaces
the existing "diag" command.

Usage: brkt aws share-logs --instance <instance_id>
--snapshot <snapshot_id> --region <region>
--bucket <bucket_name> --log-path <file_path>