Simplify the logic that handles BracketEnvironment
Move the logic that handles BracketEnvironment from
instance_config_from_values() to the callsites.  We want to default the
environment to prod when encrypting and updating.  We don't want to
default it when launching with an encrypted disk, but we do when
wrapping or launching with an unencrypted disk.  And the environment can
come from either the command line or config.  It was getting hard enough
to keep track of all of these combinations that I decided to move these
decisions up to the command-line parsing code.

instance_config_from_values() now accepts a BracketEnvironment object
instead of CLIConfig.  It's the callsite's responsibility to figure out
where the environment data comes from, and whether it should be
defaulted.

Move get_yeti_service() to instance_config_args, since that's the only
callsite.