Initialize variable at the begining for subsequent use
The variable `mv_vm_name` was being assigned after the clone VM
method. As a result, if the clone VM method hit an exception (for
e.g. due to the source VM not being found), then we were not
logging the error message properly in the except block as this
variable was unassigned.