Support launch and API token expiry
Add an --exp option to the auth command, which allows the user to
specify the lifetime of the token as a time duration.  Enable make-token
--exp when getting the launch token from the Bracket service.

Deprecate specifying token expiry in years, since the number of days in
a year can vary.

Deprecate specifying token expiry as a timestamp.  The Bracket token
service does not allow the client to specify a timestamp, so that the
server timestamp is the source of truth.  We also want expiry behavior
to be consistent for both locally-generated and server-generated tokens.