Make encryption fun again! (#692)
Add the BRKT Entertainment System which lets you play games while encrypting.
Everything is ascii based and in the terminal. The last couple of log lines from the actual cli command are displayed in the game windows so you can see what's going on and when the encryption is done.

Only some commands have the --fun flag enabled. These commands are chosen by adding the gamify decorator to their target functions.

When you choose to play a game the cli creates 2 new processes: one for the command to execute and one for the game. The commands output is piped to a file so that it can be scraped by the game and also restored after the game is done. If you exit the game before the cli command is done, it will just go back to displaying the command output as it comes along.

Added dependencies:
asciimatics (1.8.0)

Testing: Ran a bunch of different commands, both successful and unsuccessful to make sure that
the game wouldn't mess with the commands themselves