Fix incorrect thumbprint generation for public keys
Sometimes the EC keypairs generated are a byte shorter.
Either x or y could be 47 bytes instead of 48 which is
mathematically valid. But when computing thumbprint,
these need to be padded with leading zeros.
Added this change.
Generated such a keypair and uploaded it to Yeti.
Made sure the KeyId computed by yeti matches the one
computed by brkt-cli

Jira: YETI-2778 #resolve