Use our own models for unit testing
Copy the boto2 classes into a new brkt_cli.aws.model module, and use
them for unit testing.  This removes dependencies on boto2 and prepares
us for the migration to boto3.

boto3 does not have statically defined models.  It generates classes and
fields from JSON schema.  When we migrate to boto3, we'll likely need to
tweak these classes to match the objects returned by the boto3 API.