Simplify the update_ami code
Simplify the update_ami code and remove a bunch of (hopefully)
unnecessary logic.

Hardcode the Metavisor and guest root device names, instead of reading
them from the block device mappings.  We don't expect these to change
any time soon.

Consolidate the hacks that preserve volume settings, such as volume type
and IOPS, so that they don't obfuscate the rest of the code.