Paginate file list in GCP buckets
The GCP bucket list command returns the first 1000 items by default.
In the case where a bucket contains more than 1000 items, we do not
paginate thru' the entire list of files and as a result only
operate on the first 1000 entries. This especially causes problems
when you specify an encryptor bucket with more than 1000 files as
it incorrectly picks an older Metavisor image.