Improve unit testing of attach/detach volume
When detaching a volume during a unit test, remove it from the
instance's block device mapping.  Add a check for attempting to attach a
volume when the device name is already in use.  Update the
detach_volume() function signature, so that the instance ID is required.

Remove test_register_ami_hvm.  This test started failing because
_register_ami() assumes that the root device has already been removed
from the instance.  I verified that the existing tests cover the
non-legacy code path.