Turn on ENA support if Metavisor supports it
Enable ENA support on the guest instance if it's enabled on the
Metavisor instance.  Shuffle around some of the code in update_ami, to
allow us to do this earlier in the process.  If things go wrong, we'll
know sooner rather than later.  Log the "setting attribute" message at
info level for greater transparency.