Disable legacy mode
Disable the legacy code path in encrypt-ami. This code was put in to
deal with inconsistencies in block device mappings, and problems with
unmounting the root volume from the guest. I believe that this
workaround is no longer necessary with the latest HVM images. I tested
the non-legacy code path with Ubuntu 12.04-16.04, CentOS 6 and 7, Amazon
Linux, RHEL 6.9-7.3, and Cloud Foundry BOSH images. None of them
required the workaround.

Set legacy mode to False by default. Add a hidden --legacy option that
can be used to reinstate legacy mode in the field, in case this change
causes problems.  We can remove the legacy code altogether once we're
convinced that we didn't break backward compatibility.