Add option for specifying the public API CA cert
Add a --public-api-ca-cert option to all commands that talk to the
Bracket service.  This option is used for specifying the path to the
self-signed root cert when brkt-cli talks to a Customer Managed MCP.

Also add a config environment key for this option.

Add crypto.validate_cert_path() for validating a cert that's stored on
disk.  Call this function when validating the --public-api-ca-cert and
--ca-cert values.