Fix IAM role settings with boto3
After the migration to boto3, the validation method for the IAM role
as well as the instance launch method with the IAM role were failing.
Those were fixed to match the required boto3 syntax.

Also included @billybobjoeaglt change to specify a guest OS instead
of AMI ID for the wrap command.