Fix when rename VM does not work.
In case rename VM does not work, e.g. for VSAN, use the not-error-free
solution where the old template vm gets deleted, and then the new one
gets created. Thus, this can cause issues if creation of new template
fails because old one gets deleted. But, without rename, this is
the easiest fix and can be revisited if customers run into problems.