Remove timestamp deprecation
Remove deprecation warnings that are logged when the user specifies
token expiry as a timestamp (as opposed to a duration).  It's convenient
to specify that a token expires at a given date and time.  When the user
specifies a timestamp, the CLI calculates the time difference and sends
it to the token service as a value in seconds.