Preserve license info when wrapping an image
Rewrite the wrap-guest-image logic so that we modify the guest instance
instead of launching a new one.  This preserves RHEL licensing
information that's embedded in the instance and allows the user to run
yum update.

The new process is:

1) Launch the guest.
2) Stop the guest.
3) Create a Metavisor root volume based on the Metavisor AMI.
4) Move the guest root to /dev/sdf.
5) Attach the Metavisor volume at the original guest root location.
6) Start the guest.

The whole process takes about a minute.

Random related fixes:
* Add AWSService.start_instance().
* Add support for modifying user-data of an instance.  This has to be
special-cased because user-data isn't a simple string value.
* Add initialization of volume size and instance placement to
DummyAWSService, so that launch_wrapped_image() gets the values it's
expecting.
* Update unit tests to check that we launch the guest instead of
launching Metavisor.