Refactor _validate() so that it does not depend on command-line arguments
Change the signature of _validate() so that it accepts individual
values, instead of making assumptions about the structure of the
argparse Namespace object.  Multiple commands call this function, and
not all of them have the same command-line arguments.  This is in
preparation for the wrap-instance command, which will add more
disparity.