Add aws wrap-instance command
Add a wrap-instance command for wrapping an AWS instance with Metavisor.
  This logic is effectively the same thing as wrapping an image, except
that we don't launch an instance at the beginning of the process.

Move the bulk of the orchestration code from launch_wrapped_image() into
a new wrap_instance() function.  Add a new wrap-instance command that
calls wrap_instance() directly.