Save the console log when encryption doesn't start
If we hit a timeout waiting for encryption to start, save the encryptor
console log in the same way that we do when encryption fails.  This
helps the user diagnose communication issues between the CLI and
encryptor, or between encryptor and Yeti.

Add an encryption_start_timeout argument to the encrypt() function.
This allows encryption to time out immediately in unit tests.