Provide option to override encryptor instance type
The default encryptor instance type is c4.xlarge which can sometimes
be not available due to capacity issues. This argument allows a user
to override and use a different encryptor instance type to
temporarily workaround the problems. In most cases, the default
instance type would suffice (and would be used).